pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "the-puzzle-network"
version = "0.1.0"
description = "Multi-agent AI system for generating daily word games"
authors = [{name = "The Puzzle Network Team"}]
requires-python = ">=3.10"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "bandit[toml]>=1.7.0",
    "beautifulsoup4>=4.14.0",
    "build>=1.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.6.0",
    "safety>=3.0.0",
    "twine>=4.0.0"
]

[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = [
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]  # Allow unused args and assert in tests

[tool.ruff.lint.isort]
known-first-party = ["the_puzzle_network"]
force-single-line = false
lines-after-imports = 2

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient for gradual typing
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "agent_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dotenv"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bs4.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
